List all schedules in schedule indicator tooltip
