+++ packages/desktop-client/src/components/budget/util.ts
@@ -3,15 +3,13 @@ import { type CSSProperties } from 'react';
 
 import { styles } from '@actual-app/components/styles';
 import { theme } from '@actual-app/components/theme';
-import { t, type TFunction } from 'i18next';
+import { t } from 'i18next';
 
-import { type ScheduleStatusType } from 'loot-core/client/data-hooks/schedules';
 import { type useSpreadsheet } from 'loot-core/client/SpreadsheetProvider';
 import { send } from 'loot-core/platform/client/fetch';
 import * as monthUtils from 'loot-core/shared/months';
 import { type Handlers } from 'loot-core/types/handlers';
 import {
-  type ScheduleEntity,
   type CategoryEntity,
   type CategoryGroupEntity,
 } from 'loot-core/types/models';
@@ -196,52 +194,4 @@ export async function prewarmAllMonths(
   await Promise.all(
     months.map(month => prewarmMonth(budgetType, spreadsheet, month)),
   );
-}
-
-export function getScheduleStatusTooltip({
-  t,
-  schedule,
-  scheduleStatus,
-  locale,
-}: {
-  t: TFunction;
-  schedule: ScheduleEntity;
-  scheduleStatus: ScheduleStatusType;
-  locale?: Locale;
-}) {
-  const isToday = monthUtils.isCurrentDay(schedule.next_date);
-  const distanceFromNow = monthUtils.formatDistance(
-    schedule.next_date,
-    monthUtils.currentDay(),
-    locale,
-    {
-      addSuffix: true,
-    },
-  );
-  const formattedDate = monthUtils.format(schedule.next_date, 'MMMM d', locale);
-  switch (scheduleStatus) {
-    case 'missed':
-      return t(
-        'Missed {{scheduleName}} due {{distanceFromNow}} ({{formattedDate}})',
-        {
-          scheduleName: schedule.name,
-          distanceFromNow,
-          formattedDate,
-        },
-      );
-    case 'due':
-    case 'upcoming':
-      return t(
-        '{{scheduleName}} is due {{distanceFromNow}} ({{formattedDate}})',
-        {
-          scheduleName: schedule.name,
-          distanceFromNow: isToday ? t('today') : distanceFromNow,
-          formattedDate,
-        },
-      );
-    default:
-      throw new Error(
-        `Unsupported schedule status for tooltip: ${scheduleStatus}`,
-      );
-  }
 }
